Changed tests to clear out the current database instead of trying to delete/re-create it. Fixes up other test inconsistencies and more error checking on the ApplicationContext.Dispose.

+++ src/Umbraco.Tests/TestHelpers/BaseDatabaseFactoryTest.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Configuration;
 using System.Data.SqlServerCe;
 using System.IO;
+using System.Linq;
 using System.Web.Routing;
 using System.Xml;
 using NUnit.Framework;
@@ -33,6 +34,9 @@ namespace Umbraco.Tests.TestHelpers
     [TestFixture, RequiresSTA]
     public abstract class BaseDatabaseFactoryTest : BaseUmbracoApplicationTest
     {
+        private static volatile bool _firstRun = true;
+        private static readonly object Locker = new object();
+
         [SetUp]
         public override void Initialize()
         {
@@ -41,46 +45,86 @@ namespace Umbraco.Tests.TestHelpers
             var path = TestHelper.CurrentAssemblyDirectory;
             AppDomain.CurrentDomain.SetData("DataDirectory", path);
 
-            //Ensure that any database connections from a previous test is disposed. This is really just double safety as its also done in the TearDown.
-            if(ApplicationContext != null && DatabaseContext != null)
-                DatabaseContext.Database.Dispose();
-            SqlCeContextGuardian.CloseBackgroundConnection();
-
-            try
-            {
-                //Delete database file before continueing
-                string filePath = string.Concat(path, "\\UmbracoPetaPocoTests.sdf");
-                if (File.Exists(filePath))
-                {
-                    File.Delete(filePath);
-                }
-            }
-            catch (Exception)
-            {
-                //if this doesn't work we have to make sure everything is reset! otherwise
-                // well run into issues because we've already set some things up
-                TearDown();
-                throw;
-            }
-            
-            //Get the connectionstring settings from config
-            var settings = ConfigurationManager.ConnectionStrings[Core.Configuration.GlobalSettings.UmbracoConnectionName];
-            ConfigurationManager.AppSettings.Set(Core.Configuration.GlobalSettings.UmbracoConnectionName, @"datalayer=SQLCE4Umbraco.SqlCEHelper,SQLCE4Umbraco;data source=|DataDirectory|\UmbracoPetaPocoTests.sdf");
-            
-            //Create the Sql CE database
-            var engine = new SqlCeEngine(settings.ConnectionString);
-            engine.CreateDatabase();
-            
             ApplicationContext.Current = new ApplicationContext(
 				//assign the db context
 				new DatabaseContext(new DefaultDatabaseFactory()),
 				//assign the service context
 				new ServiceContext(new PetaPocoUnitOfWorkProvider(), new FileUnitOfWorkProvider(), new PublishingStrategy())) { IsReady = true };
-            
+
+            CreateDatabase();
+
+            DatabaseContext.Initialize();
+
             InitializeDatabase();
 
             //ensure the configuration matches the current version for tests
             SettingsForTests.ConfigurationStatus = UmbracoVersion.Current.ToString(3);
+        }
+
+        protected virtual bool EnsureBrandNewDatabase
+        {
+            get { return false; }
+        }
+
+        /// <summary>
+        /// Creates the SqlCe database if required
+        /// </summary>
+        protected virtual void CreateDatabase()
+        {
+            //this needs to be thread-safe
+            var isFirstRun = false;
+            if (_firstRun)
+            {
+                lock (Locker)
+                {
+                    if (_firstRun)
+                    {
+                        isFirstRun = true; //set the flag
+                        _firstRun = false;
+                    }
+                }
+            }
+
+            var path = TestHelper.CurrentAssemblyDirectory;
+
+            //Ensure that any database connections from a previous test is disposed. 
+            //This is really just double safety as its also done in the TearDown.
+            if (ApplicationContext != null && DatabaseContext != null)
+                DatabaseContext.Database.Dispose();
+            SqlCeContextGuardian.CloseBackgroundConnection();
+
+            //Get the connectionstring settings from config
+            var settings = ConfigurationManager.ConnectionStrings[Core.Configuration.GlobalSettings.UmbracoConnectionName];
+            ConfigurationManager.AppSettings.Set(Core.Configuration.GlobalSettings.UmbracoConnectionName, @"datalayer=SQLCE4Umbraco.SqlCEHelper,SQLCE4Umbraco;data source=|DataDirectory|\UmbracoPetaPocoTests.sdf");
+
+            string dbFilePath = string.Concat(path, "\\UmbracoPetaPocoTests.sdf");
+
+            if (isFirstRun || EnsureBrandNewDatabase || !File.Exists(dbFilePath))
+            {
+                try
+                {
+                    //Delete database file before continueing                    
+                    if (File.Exists(dbFilePath))
+                    {
+                        File.Delete(dbFilePath);
+                    }
+                }
+                catch (Exception)
+                {
+                    //if this doesn't work we have to make sure everything is reset! otherwise
+                    // well run into issues because we've already set some things up
+                    TearDown();
+                    throw;
+                }
+
+                //Create the Sql CE database
+                var engine = new SqlCeEngine(settings.ConnectionString);
+                engine.CreateDatabase();
+            }
+            else
+            {
+                TestHelper.ClearDatabase();
+            }            
         }
 
         /// <summary>
@@ -103,12 +147,13 @@ namespace Umbraco.Tests.TestHelpers
             base.FreezeResolution();
         }
 
+        /// <summary>
+        /// Creates the tables and data for the database
+        /// </summary>
         protected virtual void InitializeDatabase()
         {
-            //Configure the Database and Sql Syntax based on connection string set in config
-            DatabaseContext.Initialize();
             //Create the umbraco database and its base data
-            DatabaseContext.Database.CreateDatabaseSchema(false);
+            DatabaseContext.Database.CreateDatabaseSchema(false);            
         }
 
         [TearDown]
@@ -116,28 +161,36 @@ namespace Umbraco.Tests.TestHelpers
         {
             base.TearDown();
 
+            string path = TestHelper.CurrentAssemblyDirectory;
+
             SqlSyntaxContext.SqlSyntaxProvider = null;
             
 			//legacy API database connection close - because a unit test using PetaPoco db-layer can trigger the usage of SqlHelper we need to ensure that a possible connection is closed.
 			SqlCeContextGuardian.CloseBackgroundConnection();
 			
-            string path = TestHelper.CurrentAssemblyDirectory;
-            AppDomain.CurrentDomain.SetData("DataDirectory", null);
-            
-            try
+            if (EnsureBrandNewDatabase)
             {
-                string filePath = string.Concat(path, "\\UmbracoPetaPocoTests.sdf");
-                if (File.Exists(filePath))
+                try
                 {
-                    File.Delete(filePath);
+                    string filePath = string.Concat(path, "\\UmbracoPetaPocoTests.sdf");
+                    if (File.Exists(filePath))
+                    {
+                        File.Delete(filePath);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    LogHelper.Error<BaseDatabaseFactoryTest>("Could not remove the old database file", ex);
+
+                    //We will swallow this exception! That's because a sub class might require further teardown logic.
                 }
             }
-            catch (Exception ex)
+            else
             {
-                LogHelper.Error<BaseDatabaseFactoryTest>("Could not remove the old database file", ex);
-
-                //We will swallow this exception! That's because a sub class might require further teardown logic.
+                TestHelper.ClearDatabase();
             }
+           
+            AppDomain.CurrentDomain.SetData("DataDirectory", null);
                             
         }
